requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
httpx==0.27.0
pydantic==2.8.2
python-multipart==0.0.9
Pillow==10.3.0
python-dotenv==1.0.1
openai==1.43.0
requests==2.32.3
PyJWT==2.9.0
structlog==24.4.0
pytest==8.3.3
respx==0.21.1
